+++ react/features/app/middlewares.any.ts
@@ -1,4 +1,5 @@
 import '../analytics/middleware';
+import '../authentication/middleware';
 import '../av-moderation/middleware';
 import '../base/conference/middleware';
 import '../base/config/middleware';
