+++ react/features/app/middlewares.web.ts
@@ -1,4 +1,4 @@
-import '../authentication/middleware';
+import '../base/connection/middleware';
 import '../base/i18n/middleware';
 import '../base/devices/middleware';
 import '../base/media/middleware';
