+++ react/features/base/connection/actions.web.ts
@@ -1,52 +1,67 @@
+// @ts-expect-error
+import { jitsiLocalStorage } from '@jitsi/js-utils';
+
 import { IStore } from '../../app/types';
+import { getCustomerDetails } from '../../jaas/actions.any';
+import { getJaasJWT, isVpaasMeeting } from '../../jaas/functions';
 import { showWarningNotification } from '../../notifications/actions';
 import { NOTIFICATION_TIMEOUT_TYPE } from '../../notifications/constants';
 import { stopLocalVideoRecording } from '../../recording/actions.any';
 import LocalRecordingManager from '../../recording/components/Recording/LocalRecordingManager.web';
-import { configureInitialDevices } from '../devices/actions';
-import { getBackendSafeRoomName } from '../util/uri';
+import { setJWT } from '../jwt/actions';
 
-export {
-    connectionDisconnected,
-    connectionEstablished,
-    connectionFailed,
-    setLocationURL
-} from './actions.any';
-import logger from './logger';
+import { _connectInternal } from './actions.any';
 
 export * from './actions.any';
 
 /**
  * Opens new connection.
  *
- * @returns {Promise<JitsiConnection>}
+ * @param {string} [id] - The XMPP user's ID (e.g. {@code user@server.com}).
+ * @param {string} [password] - The XMPP user's password.
+ * @returns {Function}
  */
-export function connect() {
+export function connect(id?: string, password?: string) {
     return (dispatch: IStore['dispatch'], getState: IStore['getState']) => {
-        const room = getBackendSafeRoomName(getState()['features/base/conference'].room);
+        const state = getState();
+        const { jwt } = state['features/base/jwt'];
+        const { iAmRecorder, iAmSipGateway } = state['features/base/config'];
 
-        // XXX For web based version we use conference initialization logic
-        // from the old app (at the moment of writing).
-        return dispatch(configureInitialDevices()).then(
-            () => APP.conference.init({
-                roomName: room
-            }).catch((error: Error) => {
-                APP.API.notifyConferenceLeft(APP.conference.roomName);
-                logger.error(error);
-            }));
+        if (!iAmRecorder && !iAmSipGateway && isVpaasMeeting(state)) {
+            return dispatch(getCustomerDetails())
+                .then(() => {
+                    if (!jwt) {
+                        return getJaasJWT(state);
+                    }
+                })
+                .then(j => j && dispatch(setJWT(j)))
+                .then(() => _connectInternal(id, password));
+        }
+
+        // used by jibri
+        const usernameOverride = jitsiLocalStorage.getItem('xmpp_username_override');
+        const passwordOverride = jitsiLocalStorage.getItem('xmpp_password_override');
+
+        if (usernameOverride && usernameOverride.length > 0) {
+            id = usernameOverride; // eslint-disable-line no-param-reassign
+        }
+        if (passwordOverride && passwordOverride.length > 0) {
+            password = passwordOverride; // eslint-disable-line no-param-reassign
+        }
+
+        return dispatch(_connectInternal(id, password));
     };
 }
 
 /**
  * Closes connection.
  *
- * @param {boolean} [requestFeedback] - Whether or not to attempt showing a
+ * @param {boolean} [requestFeedback] - Whether to attempt showing a
  * request for call feedback.
  * @returns {Function}
  */
-export function disconnect(requestFeedback = false) {
-    // XXX For web based version we use conference hanging up logic from the old
-    // app.
+export function hangup(requestFeedback = false) {
+    // XXX For web based version we use conference hanging up logic from the old app.
     return async (dispatch: IStore['dispatch']) => {
         if (LocalRecordingManager.isRecordingLocally()) {
             dispatch(stopLocalVideoRecording());
