+++ react/features/conference/actions.web.ts
@@ -1,9 +1,12 @@
 import { IStore } from '../app/types';
+import { configureInitialDevices } from '../base/devices/actions.web';
 import { getParticipantDisplayName } from '../base/participants/functions';
+import { getBackendSafeRoomName } from '../base/util/uri';
 import { showNotification } from '../notifications/actions';
 import { NOTIFICATION_TIMEOUT_TYPE, NOTIFICATION_TYPE } from '../notifications/constants';
 
 import { DISMISS_CALENDAR_NOTIFICATION } from './actionTypes';
+import logger from './logger';
 
 /**
  * Notify that we've been kicked out of the conference.
@@ -43,5 +46,26 @@ export function notifyKickedOut(participant: any, _?: Function) {
 export function dismissCalendarNotification() {
     return {
         type: DISMISS_CALENDAR_NOTIFICATION
+    };
+}
+
+/**
+ * Init.
+ *
+ * @returns {Promise<JitsiConnection>}
+ */
+export function init() {
+    return (dispatch: IStore['dispatch'], getState: IStore['getState']) => {
+        const room = getBackendSafeRoomName(getState()['features/base/conference'].room);
+
+        // XXX For web based version we use conference initialization logic
+        // from the old app (at the moment of writing).
+        return dispatch(configureInitialDevices()).then(
+            () => APP.conference.init({
+                roomName: room
+            }).catch((error: Error) => {
+                APP.API.notifyConferenceLeft(APP.conference.roomName);
+                logger.error(error);
+            }));
     };
 }
