+++ react/features/settings/actions.native.ts
@@ -1,0 +1,10 @@
+/**
+ * Opens {@code LogoutDialog}.
+ *
+ * @param {Function} onLogout - The event in {@code LogoutDialog} that should be
+ *  enabled on click.
+ * @returns {Function}
+ */
+export function openLogoutDialog(onLogout: Function) { // eslint-disable-line @typescript-eslint/no-unused-vars
+    // this is available only for web at the moment
+}
